Fixed memory leak issues when using g_variant_get() and g_variant_iter_loop()
This patch fixes memory leak issues when using g_variant_get() API.
As per GLib documentation:
If we extract string as "&s" or object patch as "&o" then we are not
required to free the memory otherwise we need to free the memory if we
extract string as "s" or object_path as "o".
Also when using g_variant_iter_loop() and if we break from it before loop ends,
then we need to free the variant extracted from it, So changed obejct_path to "&o".

Change-Id: I42845944dea449112f0f2c959464c27ca27af979
Signed-off-by: Nishant Chaprana <n.chaprana@samsung.com>