Fixed chance of crash when sending notification to client
Description: notification variable should be local variable,
instead of a pointer because when it is taken as a pointer and
g_try_malloc0() fails to allocate memory, in that case crash
will occur by dereferencing "noti" pointer.
Also if we add a null check for noti allocation failure, then
we would fail to send notification to application, which is not
correct operation.
So noti should a local variable instead of pointer.

Change-Id: I83f29a259e57cafb71779e0485f320a26484ce9d
Signed-off-by: Nishant Chaprana <n.chaprana@samsung.com>